requirements.txt
Flask
Flask-Cors
flask_marshmallow
SQLAlchemy
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
flasgger
marshmallow-sqlalchemy